+++ planner/core/src/main/java/eu/seaclouds/platform/planner/core/application/topology/modifier/relation/TopologFacadeyModifierApplicator.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.seaclouds.platform.planner.core.application.topology.modifier;
+package eu.seaclouds.platform.planner.core.application.topology.modifier.relation;
 
 
 import com.google.common.collect.ImmutableList;
